issue #297: updated comments and doc blocks, improved some methods

+++ system/classes/db.php
@@ -1,62 +1,84 @@
 <?php
-namespace YAWK {
+namespace YAWK
+{
     /**
-     * <b>Database class db</b> <i>extends \mysqli</i>
+     * <b>Database class - connect to mysqli and return connection object</b>
+     * <p>This class establish the database connection if none already exists.
+     * It serves some handy methods as well, like quote data, get and delete tables,
+     * import .sql files, as well as typical select and query methods. </p>
+     *
+     * @author     Daniel Retzl <danielretzl@gmail.com>
+     * @copyright  2012-2018 Daniel Retzl yawk.io
+     * @license    https://opensource.org/licenses/MIT
+     * @version    1.0.0
+     * @link       http://yawk.io
+     * @annotation Mysqli database class
      */
-    class db extends \mysqli {
+    class db {
+
+        /** @var array $config mysql configuration (host, username, database etc...) */
         private $config;
+        /** @var object $connection holds the mysqli database connection */
         private $connection;
 
         /**
-         * db constructor.
+         * db constructor - Include the database config file
          */
         public function __construct()
-        {
+        {   // include config array
             require_once ("dbconfig.php");
         }
 
         /**
          * Connect to the database
          *
-         * @return object | bool false on failure / mysqli MySQLi object instance on success
+         * @return object|bool false on failure / mysqli MySQLi object instance on success
          */
         public function connect()
         {
-            // Try and connect to the database
+            // if connection is not set
             if (!isset($this->connection))
             {
-                $this->connection = new \mysqli($this->config['server'], $this->config['username'], $this->config['password'], $this->config['dbname'], $this->config['port']);
-
-                if (!$this->connection)
+                // create new database object
+                if ($this->connection = new \mysqli(
+                        $this->config['server'],
+                        $this->config['username'],
+                        $this->config['password'],
+                        $this->config['dbname'],
+                        $this->config['port']))
                 {
-                    return false;
+                    // connection established successfully
+                    return $this->connection;
                 }
                 else
-                    {
-                        return $this->connection;
+                    {   // failed to connect to database
+                        // (wrong credentials?)
+                        return false;
                     }
             }
             else
-                {
+                {   // connection already established...
                     return $this->connection;
                 }
         }
 
         /**
-         * query database
-         *
-         * @param $query | the query string
-         * @return mixed The result of the mysqli::query() function
+         * Execute any sql query
+         * @param string $query the sql query string
+         * @return mixed The mysqli result
          */
         public function query($query)
         {
-            // connect to the database
+            // connect to database
             if ($this->connection = $this->connect())
-            {   // if connection is successful
-                // replace {} in str for db prefix
+            {
+                // if connection is successful
+
+                // replace {} in str to add table prefix
                 $query = str_replace("}", "", $query);
                 $query = str_replace("{", $this->config['prefix'], $query);
-                // query the database
+
+                // query database
                 return $this->connection->query($query);
             }
             else
@@ -66,49 +88,72 @@ namespace YAWK {
         }
 
         /**
-         * Fetch rows from the database (SELECT query)
-         *
-         * @param $query | The query string
-         * @return array | bool on failure
+         * Fetch rows from database (SELECT query)
+         * @param string $query the sql query string
+         * @return array|bool on failure
          */
         public function select($query)
         {
+            // init result rows array
             $rows = array();
+            // query database
             $result = $this->query($query);
-            if ($result === false) {
+            // check if result is false
+            if ($result === false)
+            {
                 // $error = $this->error();
                 //\YAWK\sys::setSyslog($db, 5, "$error", 0, 0, 0, 0);
                 return false;
             }
-            while ($row = $result->fetch_assoc()) {
-                $rows[] = $row;
+            else
+            {   // fetch associative result
+                while ($row = $result->fetch_assoc())
+                {   // fill array
+                    $rows[] = $row;
+                }
             }
-            return $rows;
+            // check if result rows array is set
+            if (isset($rows) && (!empty($rows) && (is_array($rows))))
+            {   // ok, return result rows
+                return $rows;
+            }
+            else
+                {   // no result set
+                    return false;
+                }
         }
 
         /**
-         * Fetch the last error from the database
-         *
-         * @return string Database error message
+         * Fetch last error from the database
+         * @return string database error
          */
         public function error()
         {
+            // connect to database
             $this->connection = $this->connect();
+            // return latest error
             return $this->connection->error;
         }
 
         /**
-         * Quote and escape value for use in a database query
-         *
-         * @param string $value The value to be quoted and escaped
-         * @return string The quoted and escaped string
+         * Quote and escape value for use in a database query         *
+         * @param string $value the value to be quoted and escaped
+         * @return string the quoted and escaped string
          */
         public function quote($value)
         {
+            // connect to database
             $this->connection = $this->connect();
+            // escape and return string
             return $this->connection->real_escape_string($value);
         }
 
+        /**
+         * Import data from .sql file into database
+         * @param $sqlfile
+         * @param $lang
+         * @return bool
+         */
         public function import($sqlfile, $lang)
         {
             // http://stackoverflow.com/questions/19751354/how-to-import-sql-file-in-mysql-database-using-php
@@ -180,34 +225,51 @@ namespace YAWK {
         } // ./ import
 
 
+        /**
+         * Delete a whole database (including all tables)
+         * @param $database
+         */
         public function deleteDatabase($database)
         {
+            // get all tables as array
             $result = $this->query("SHOW TABLES IN `$database`");
-            // $result = mysqli_query($db, "SHOW TABLES IN `$databaseName`");
-            while ($table = mysqli_fetch_array($result))
+            // check if result is set
+            if (isset($result) && (!empty($result)))
             {
-                $tableName = $table[0];
-                if ($this->query("TRUNCATE TABLE `$database`.`$tableName`"))
-                {
-                    echo "$tableName was cleared <br>";
-                }
-                else
-                    {
-                        echo mysqli_errno($this). ' ' . mysqli_error($this).'<br>';
+                // walk through result array
+                while ($table = mysqli_fetch_array($result))
+                {   // store tablename in var
+                    $tableName = $table[0];
+                    // try to delete table
+                    if ($this->query("TRUNCATE TABLE `$database`.`$tableName`"))
+                    {   // output(?)
+                        echo "$tableName was cleared <br>";
                     }
+                    else
+                    {   // output error
+                        echo $this->error().'<br>';
+                    }
+                }
             }
         }
 
+        /**
+         * Get all tables from a database and return as array
+         * @return array tableList
+         */
         public function get_tables()
         {
+            // set tableList array
             $tableList = array();
+            // get tables from database
             $res = $this->query("SHOW TABLES");
+            // walk through result
             while($row = mysqli_fetch_array($res))
-            {
+            {   // fill array
                 $tableList[] = $row[0];
             }
             return $tableList;
         }
 
-    } // ./dbclass
+    } // EOF ./dbclass
 }// ./namespace
